[Silver II] Title: 알고리즘 수업 - 너비 우선 탐색 4, Time: 76 ms, Memory: 9108 KB -BaekjoonHub

+++ 백준/Silver/24447. 알고리즘 수업 － 너비 우선 탐색 4/알고리즘 수업 － 너비 우선 탐색 4.cc
@@ -8,7 +8,7 @@ int main() {
     
     int N, M, R; cin >> N >> M >> R;
     vector<int> G[N+1];
-    queue<pair<int,long long>> Q;
+    queue<pair<int,int>> Q;
     int visit[N+1]{0};
     long long result = 0, t = 0;
 
@@ -23,7 +23,7 @@ int main() {
 
     while(!Q.empty()) {
         int n = Q.front().first;
-        long long d = Q.front().second;
+        int d = Q.front().second;
 
         t++;
         result += d*t;
